# shared/openapi.yaml
openapi: 3.0.3
info:
  title: Credit Intel Prototype API
  version: 0.1.0
  description: Shared API contract for frontend, Go backend, and Python AI engine.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service healthy
  /cases:
    post:
      summary: Create a new credit case
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCaseInput"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCase"
        "400":
          description: Invalid input
  /cases/{id}:
    get:
      summary: Get case by ID
      operationId: getCase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCase"
        "404":
          description: Case not found
  /cases/{id}/upload:
    post:
      summary: Upload files for a case
      operationId: uploadFiles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Files uploaded successfully
        "404":
          description: Case not found
  /cases/{id}/analyze:
    post:
      summary: Trigger AI analysis for a case
      operationId: analyzeCase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Analysis started
        "404":
          description: Case not found
  /cases/{id}/cam:
    get:
      summary: Get CAM result for a case
      operationId: getCAM
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CAM result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CAMResult"
        "404":
          description: CAM not yet generated or case not found
  /cases/{id}/cam/download:
    get:
      summary: Download CAM document (DOCX or Markdown)
      operationId: downloadCAM
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CAM document file
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "404":
          description: CAM not yet generated or case not found
components:
  schemas:
    CreditCaseInput:
      type: object
      required:
        - company_name
        - sector
      properties:
        company_name:
          type: string
        cin_optional:
          type: string
        sector:
          type: string
        promoter_names:
          type: array
          items:
            type: string
        officer_notes:
          type: string
    CreditCase:
      type: object
      properties:
        case_id:
          type: string
        company_name:
          type: string
        cin_optional:
          type: string
        sector:
          type: string
        promoter_names:
          type: array
          items:
            type: string
        uploaded_files:
          type: array
          items:
            $ref: "#/components/schemas/UploadedFile"
        officer_notes:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [created, processing, ready, failed]
        extracted_facts:
          $ref: "#/components/schemas/ExtractedFacts"
        risk_flags:
          $ref: "#/components/schemas/RiskFlags"
        cam_result:
          $ref: "#/components/schemas/CAMResult"
        score_result:
          $ref: "#/components/schemas/ScoreResult"
        officer_note_signals:
          $ref: "#/components/schemas/OfficerNoteSignals"
    UploadedFile:
      type: object
      properties:
        file_name:
          type: string
        file_path:
          type: string
        doc_type:
          type: string
        uploaded_at:
          type: string
          format: date-time
    ExtractedFacts:
      type: object
      properties:
        revenue:
          type: number
        EBITDA:
          type: number
        PAT:
          type: number
        total_debt:
          type: number
        current_ratio:
          type: number
        dscr:
          type: number
        working_capital:
          type: number
        contingent_liabilities:
          type: array
          items:
            type: string
        related_party_transactions:
          type: array
          items:
            type: string
        auditor_remarks:
          type: array
          items:
            type: string
        extracted_entities:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              role:
                type: string
              source_ref:
                type: string
        document_sources:
          type: array
          items:
            type: object
            properties:
              doc_id:
                type: string
              file_name:
                type: string
              section:
                type: string
              page_ref:
                type: string
    RiskFlag:
      type: object
      required:
        - flag_id
        - flag_type
        - severity
        - description
      properties:
        flag_id:
          type: string
        flag_type:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        evidence_refs:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        impact_on_score:
          type: string
    RiskFlags:
      type: array
      items:
        $ref: "#/components/schemas/RiskFlag"
    CAMResult:
      type: object
      required:
        - case_id
        - final_decision
        - generated_at
      properties:
        case_id:
          type: string
        final_decision:
          type: string
          enum: [approve, approve_with_conditions, decline, manual_review]
        recommended_limit:
          type: number
        recommended_roi:
          type: number
        overall_score:
          type: number
        score_breakdown:
          type: object
          additionalProperties:
            type: number
        key_reasons:
          type: array
          items:
            type: string
        evidence_summary:
          type: string
        cam_doc_path:
          type: string
        generated_at:
          type: string
          format: date-time
    OfficerNoteSignalDetail:
      type: object
      properties:
        score:
          type: number
        explanations:
          type: array
          items:
            type: string
    OfficerNoteSignals:
      type: object
      properties:
        capacity_utilization:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        management_quality:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        operational_health:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        collection_risk:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        site_visit_risk:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        promoter_behavior_score:
          $ref: "#/components/schemas/OfficerNoteSignalDetail"
        composite_score:
          type: number
        all_explanations:
          type: array
          items:
            type: string
    ScoreResult:
      type: object
      properties:
        overall_score:
          type: number
        score_breakdown:
          type: object
          additionalProperties:
            type: number
        decision:
          type: string
          enum: [approve, review, reject]
        decision_explanation:
          type: string
        recommended_limit:
          type: number
        recommended_roi:
          type: number
        reasons:
          type: array
          items:
            type: string
        hard_override_applied:
          type: boolean
        hard_override_reason:
          type: string
        officer_note_signals:
          $ref: "#/components/schemas/OfficerNoteSignals"